changes
changes

+++ Class A6/ActivityLifeCycle/ActivityLifeCycle/MainActivity.cs
@@ -16,43 +16,43 @@ namespace ActivityLifeCycle
 		{
 			base.OnCreate (bundle);
 			SetContentView (Resource.Layout.Main);
-			Toast.MakeText (this,"On Create",ToastLength.Short).Show ();
+			Toast.MakeText (this,"On Create",ToastLength.Long).Show ();
 		}
 
 		protected override void OnStart()
 		{
 			base.OnStart ();
-			Toast.MakeText (this,"On Start",ToastLength.Short).Show ();
+			Toast.MakeText (this,"On Start",ToastLength.Long).Show ();
 		}
 
 		protected override void OnResume()
 		{
 			base.OnResume ();
-			Toast.MakeText (this,"On Resume",ToastLength.Short).Show ();
+			Toast.MakeText (this,"On Resume",ToastLength.Long).Show ();
 		}
 
 		protected override void OnRestart()
 		{
 			base.OnRestart ();
-			Toast.MakeText (this, "On Restart", ToastLength.Short).Show ();
+			Toast.MakeText (this, "On Restart", ToastLength.Long).Show ();
 		}
 
 		protected override void OnPause()
 		{
 			base.OnPause ();
-			Toast.MakeText (this, "On Pause", ToastLength.Short).Show ();
+			Toast.MakeText (this, "On Pause", ToastLength.Long).Show ();
 		}
 
 		protected override void OnStop()
 		{
 			base.OnStop ();
-			Toast.MakeText (this, "On Stop", ToastLength.Short).Show ();
+			Toast.MakeText (this, "On Stop", ToastLength.Long).Show ();
 		}
 
 		protected override void OnDestroy()
 		{
 			base.OnDestroy ();
-			Toast.MakeText (this, "On Destroy", ToastLength.Short).Show ();
+			Toast.MakeText (this, "On Destroy", ToastLength.Long).Show ();
 		}
 
 	}
